+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Add `EveOnline::ESI::Models::StargateDestination`
 * Add `EveOnline::ESI::Models::Stargate`
 * Add `EveOnline::ESI::UniverseStargate`
+* Add `EveOnline::ESI::Models::Star`
 
 ## v0.19.0